options: move CoupledOptions to root dir
