+++ test/adapt/test_recovery.py
@@ -5,7 +5,7 @@ import numpy as np
 
 from adapt_utils.adapt.recovery import *
 from adapt_utils.swe.utils import recover_vorticity
-from adapt_utils.unsteady.options import *
+from adapt_utils.options import CoupledOptions
 
 
 def get_mesh(dim, n):
