+++ unsteady/test_cases/cosine_prescribed_velocity/options.py
@@ -1,6 +1,6 @@
 from firedrake import *
 
-from adapt_utils.unsteady.options import CoupledOptions
+from adapt_utils.options import CoupledOptions
 
 
 __all__ = ["CosinePrescribedVelocityOptions"]
